refactor(T1d): impl-split SFHologramHelperService across 2 .cpp [BUILD-PASSED]
Final >2k file. Pure impl-split (one class, 2 .cpp). SFHologramHelperService.cpp 2,144 -> 1,161;
_Part2.cpp 1,026. Anon-namespace helper RefreshHologramVisibility stays in core with its sole user.
Zero behavior change. Shipping PackagePlugin clean (ExitCode=0). Solo lane.
